+++ lib/app.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'join_create.dart';
+
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(home: HomePage());
+  }
+}
