+++ pubspec.yaml
@@ -16,6 +16,8 @@ dependencies:
   flutter:
     sdk: flutter
 
+  frideos: ^0.4.0+2
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
